Pass snippets as argument to engine
