+++ packages/cursorless-engine/src/actions/Actions.ts
@@ -69,7 +69,7 @@ import { Decrement, Increment } from "./incrementDecrement";
 export class Actions implements ActionRecord {
   constructor(
     private treeSitter: TreeSitter,
-    private snippets: Snippets,
+    private snippets: Snippets | undefined,
     private rangeUpdater: RangeUpdater,
     private modifierStageFactory: ModifierStageFactory,
   ) {}
