+++ packages/cursorless-engine/src/index.ts
@@ -12,3 +12,6 @@ export * from "./CommandHistory";
 export * from "./CommandHistoryAnalyzer";
 export * from "./util/grammarHelpers";
 export * from "./ScopeTestRecorder";
+export * from "./core/Snippets";
+export * from "./core/mergeSnippets";
+export * from "./util/object";
