+++ packages/cursorless-vscode/src/extension.ts
@@ -51,6 +51,7 @@ import {
 import { StatusBarItem } from "./StatusBarItem";
 import { storedTargetHighlighter } from "./storedTargetHighlighter";
 import { vscodeApi } from "./vscodeApi";
+import { VscodeSnippets } from "./VscodeSnippets";
 
 /**
  * Extension entrypoint called by VSCode on Cursorless startup.
@@ -83,12 +84,14 @@ export async function activate(
 
   const treeSitter: TreeSitter = createTreeSitter(parseTreeApi);
 
+  const snippets = new VscodeSnippets(normalizedIde);
+  void snippets.init();
+
   const {
     commandApi,
     storedTargets,
     hatTokenMap,
     scopeProvider,
-    snippets,
     injectIde,
     runIntegrationTests,
     addCommandRunnerDecorator,
@@ -99,6 +102,7 @@ export async function activate(
     hats,
     commandServerApi,
     fileSystem,
+    snippets,
   );
 
   addCommandRunnerDecorator(
